+++ src/main/java/br/com/rogerio/Bank.java
@@ -1,0 +1,9 @@
+package br.com.rogerio;
+
+public class Bank {
+
+	public Money reduce(Expression source, String to) {
+		return Money.dollar(10);
+	}
+
+}