+++ src/main/java/br/com/rogerio/Expression.java
@@ -1,0 +1,6 @@
+package br.com.rogerio;
+
+interface Expression {
+
+Expression plus(Money added);
+}