+++ src/main/java/br/com/rogerio/Money.java
@@ -1,6 +1,6 @@
 package br.com.rogerio;
 
-public class Money {
+public class Money implements Expression {
 
 	protected String currency;
 
@@ -25,10 +25,6 @@ public class Money {
 				&& currency().equals(money.currency());
 	}
 	
-//	Money times(int multipler) {
-//		return null;
-//	}
-	
 	public Money times(int multiplier) {
 		return new Money(amount * multiplier, currency);
 	}
@@ -47,5 +43,8 @@ public class Money {
 
 	public static Money franc(int amount) {
 		return new Money(amount, "CHF");
+	}
+	public Money plus(Money addend) {
+		return new Money(amount + addend.amount, currency);
 	}
 }